Fetch xml from REST to local files. Redirect http request to read from local files in test env.
XML is exported for an extension context. The mock structure is going to change to entity based mock in a next commit, instead of xml based mock that is implemented in this commit. I would like to keep the xml fetching functionality since I found it handy to work with.